explicit casts

+++ IPlug/IPlugVST.cpp
@@ -199,7 +199,7 @@ EHost IPlugVST::GetHost()
 
     if (CSTR_NOT_EMPTY(productStr))
     {
-      int decVer = mHostCallback(&mAEffect, audioMasterGetVendorVersion, 0, 0, 0, 0.0f);
+      int decVer = (int) mHostCallback(&mAEffect, audioMasterGetVendorVersion, 0, 0, 0, 0.0f);
       int ver = decVer / 10000;
       int rmaj = (decVer - 10000 * ver) / 100;
       int rmin = (decVer - 10000 * ver - 100 * rmaj);
@@ -451,7 +451,7 @@ VstIntPtr VSTCALLBACK IPlugVST::VSTDispatcher(AEffect *pEffect, VstInt32 opCode,
     }
     case effSetBlockSize:
     {
-      _this->SetBlockSize(value);
+      _this->SetBlockSize((int) value);
       _this->Reset();
       return 0;
     }
@@ -551,7 +551,7 @@ VstIntPtr VSTCALLBACK IPlugVST::VSTDispatcher(AEffect *pEffect, VstInt32 opCode,
       {
         bool isBank = (!idx);
         ByteChunk* pChunk = (isBank ? &(_this->mBankState) : &(_this->mState));
-        pChunk->Resize(value);
+        pChunk->Resize((int) value);
         memcpy(pChunk->GetBytes(), ptr, value);
         int pos = 0;
         int iplugVer = _this->GetIPlugVerFromChunk(pChunk, &pos);
@@ -785,7 +785,7 @@ VstIntPtr VSTCALLBACK IPlugVST::VSTDispatcher(AEffect *pEffect, VstInt32 opCode,
       {
         if (value >= 0 && value < _this->NParams())
         {
-          if (_this->GetParam(value)->Type() != IParam::kTypeDouble)
+          if (_this->GetParam((int) value)->Type() != IParam::kTypeDouble)
           {
             return 0xbeef;
           }
